remove account wizard duplicated sass and component files, split DateOfBirth into separate component in trader, change Dropdown to autocomplete in address details for state
