+++ packages/core/src/sass/details-form.scss
@@ -20,6 +20,14 @@
         &__picker {
             margin-top: 8px;
         }
+        .calendar__cell {
+            &--disabled {
+                color: var(--text-disabled) !important;
+            }
+            &--other:not(.calendar__cell--today) {
+                color: var(--text-prominent);
+            }
+        }
     }
     &__elements {
         flex-grow: 1;
@@ -49,9 +57,5 @@
         font-size: var(--text-size-xxs);
         margin: 1.6rem 0;
         color: var(--text-general);
-    }
-    .address-state__fieldset {
-        margin-bottom: 3.2rem;
-        margin-top: -8px;
     }
 }
