+++ packages/trader/src/App/Components/Form/DateOfBirth/index.js
@@ -1,0 +1,1 @@
+export default from './date-of-birth-picker.jsx';
