--- packages/trader/src/sass/account-wizard.scss
@@ -1,197 +1,0 @@
-@import 'app/_common/base/index';
-
-.dc-modal__container_real-account-signup-modal {
-    display: flex;
-    flex-direction: column;
-    transition: all 0.3s ease-in-out;
-
-    &--success {
-        max-height: 380px;
-        border-radius: $BORDER_RADIUS;
-        overflow: hidden;
-        max-width: 478px;
-    }
-    .dc-modal-header__title {
-        text-transform: none;
-        border-top-left-radius: 1em;
-    }
-    &--error {
-        .account-wizard--error {
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            flex-direction: column;
-            height: 100%;
-            width: 100%;
-        }
-        & > svg {
-            width: 128px;
-            height: 128px;
-            align-self: center;
-            margin: 2em 0 0;
-        }
-        h1 {
-            @include typeface(--title-center-bold-black);
-            text-transform: none;
-            color: var(--text-general);
-            margin: 1.6rem;
-        }
-        p {
-            @include typeface(--paragraph-center-normal-black);
-            text-transform: none;
-            color: var(--text-prominent);
-            width: 80%;
-        }
-        a {
-            @include typeface(--paragraph-center-bold-black);
-            text-transform: none;
-            text-decoration: none;
-            color: var(--text-colored-background);
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            width: 12em;
-            align-self: center;
-        }
-    }
-}
-
-.dc-modal-header__real-account-signup-modal {
-    border-bottom: 1px solid var(--general-section-1);
-}
-
-.account-wizard {
-    display: flex;
-    flex-direction: column;
-    flex: 1 1 0; /* fix for Safari */
-    width: 100%;
-
-    h1 {
-        @include typeface(--title-center-bold-black);
-        color: var(--text-prominent);
-        text-transform: none;
-        margin-top: 3rem;
-    }
-    h2 {
-        @include typeface(--title-center-bold-black);
-        text-transform: none;
-        margin-top: 2.4rem;
-        color: var(--text-less-prominent);
-    }
-    h3 {
-        @include typeface(--small-center-normal-black);
-        text-transform: none;
-        color: var(--text-general);
-        margin-bottom: 1.6rem;
-    }
-    form {
-        display: flex;
-        flex: 1 1 0; /* fix for Safari */
-        width: 100%;
-        padding: 0 2.4rem 3.2rem;
-    }
-    &__body {
-        display: flex;
-        flex: 1 1 0; /* fix for Safari */
-        width: 100%;
-    }
-    &__set-currency {
-        p {
-            margin-top: 2.4rem;
-            @include typeface(--paragraph-center-normal-black);
-            text-transform: initial;
-            color: var(--text-prominent);
-        }
-        h2 {
-            @include typeface(--title-center-bold-black);
-            color: var(--text-prominent);
-        }
-    }
-    &__btn-area {
-        display: flex;
-        align-items: center;
-        justify-content: flex-end;
-        align-self: flex-end;
-        padding-top: 2.4em;
-        width: 100%;
-
-        .btn:not(:last-child) {
-            margin: 0 1em;
-        }
-        .btn:last-child {
-            margin: 0;
-        }
-        &--absolute {
-            bottom: 0;
-            position: absolute;
-            left: 0;
-            padding: 2.4em;
-            border-top: 1px solid var(--general-section-1);
-        }
-        &--center {
-            justify-content: center;
-            width: 100%;
-
-            & .btn {
-                min-width: 9em;
-            }
-        }
-        &--error {
-            @include typeface(--paragraph-left-normal-black);
-            text-transform: none;
-            color: var(--brand-red-coral) !important;
-        }
-    }
-    .checkbox {
-        align-items: center;
-        margin: 1rem 0 0;
-
-        &__box--active {
-            background-color: var(--brand-red-coral);
-        }
-        &__box--active:after {
-            font-size: 1.3rem;
-            line-height: 20px;
-        }
-        &__label {
-            @include typeface(--paragraph-left-normal-black);
-            text-transform: none;
-        }
-    }
-    &--disabled {
-        position: relative;
-
-        h1, h2, h3, p {
-            color: var(--text-disabled);
-        }
-        .currency-list__item {
-            pointer-events: none;
-
-            svg {
-                opacity: 0.08;
-            }
-            .label {
-                color: var(--text-disabled);
-            }
-        }
-        &-message {
-            position: absolute;
-            top: 0;
-            left: 0;
-            width: 100%;
-            height: 100%;
-            z-index: 99;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-
-            p {
-                color: var(--text-general);
-                font-size: var(--text-size-s);
-                line-height: 1.5;
-                width: 78%;
-                text-align: center;
-            }
-        }
-    }
-}
