--- packages/trader/src/sass/add-or-manage.scss
@@ -1,90 +1,0 @@
-@import 'app/_common/base/index';
-
-.add-or-manage {
-    flex-direction: row; /* fix for Safari */
-    flex-wrap: wrap;
-
-    form {
-        flex-direction: column;
-    }
-}
-
-.success-change {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    flex-grow: 1;
-
-    &__icon-area {
-        display: flex;
-        align-items: center;
-        width: 60%;
-        justify-content: space-between;
-        position: relative;
-        margin-top: 3rem;
-
-        .bottom-right-overlay {
-            position: absolute;
-            bottom: 0;
-            right: 0rem;
-            width: 3rem;
-            height: 3rem;
-        }
-        svg.inline-icon {
-            width: 90px;
-            height: 120px;
-        }
-        svg {
-            display: block;
-        }
-        &--xlarge, &--large {
-            width: auto;
-
-            .bottom-right-overlay {
-                width: 48px;
-                height: 48px;
-                right: 0;
-            }
-        }
-        &--xlarge {
-            svg:first-child {
-                width: 128px;
-                height: 128px;
-            }
-        }
-        &--large {
-            svg.inline-icon {
-                width: 120px;
-                height: 120px;
-            }
-        }
-    }
-    &__body-area {
-        flex-grow: 1;
-
-        h2 {
-            @include typeface(--title-center-bold-black);
-            margin-top: 2rem;
-            color: var(--text-general);
-        }
-        p {
-            @include typeface(--paragraph-left-normal-black);
-            padding: 1rem;
-            text-align: center;
-            text-transform: none;
-            color: var(--text-general);
-        }
-    }
-    &__btn-area {
-        margin: 3rem 0;
-        width: 100%;
-        display: flex;
-        justify-content: center;
-
-        & .btn {
-            margin: 0 2rem;
-            height: 4rem;
-        }
-    }
-}
