--- packages/trader/src/sass/change-account.scss
@@ -1,11 +1,0 @@
-@import 'app/_common/base/index';
-
-.change-currency {
-    background-color: var(--general-section-1);
-    border-bottom-left-radius: $BORDER_RADIUS;
-    border-bottom-right-radius: $BORDER_RADIUS;
-
-    h1 {
-        text-transform: none;
-    }
-}
