--- packages/trader/src/sass/currency-select-radio.scss
@@ -1,81 +1,0 @@
-@import 'app/_common/base/index';
-
-.currency-list {
-    width: 100%;
-
-    form {
-        width: 100%;
-    }
-    &__items {
-        display: flex;
-        justify-content: space-around;
-    }
-    &__item {
-        align-items: center;
-        border-radius: 8px;
-        display: inline-flex;
-        flex-direction: column;
-        height: 128px;
-        justify-content: space-around;
-        margin: 1.6rem 2rem 0;
-        width: 128px;
-
-        &:hover:not(&--selected) {
-            cursor: pointer;
-            border: 1px solid var(--border-hover);
-        }
-        &--current {
-            pointer-events: none;
-
-            svg {
-                opacity: 0.08;
-            }
-            .label {
-                color: var(--text-disabled);
-            }
-        }
-        &--selected {
-            border: 2px solid var(--border-active);
-
-            & p.label {
-                @include typeface(--paragraph-center-bold-black);
-                color: var(--text-prominent);
-            }
-        }
-        & > div {
-            display: flex;
-            flex-direction: column;
-            justify-content: space-between;
-            align-items: center;
-        }
-        svg {
-            height: 48px;
-            width: 48px;
-            margin-bottom: 1em;
-        }
-        p {
-            text-align: center;
-            @include typeface(--paragraph-center-normal-black);
-            color: var(--text-general);
-        }
-    }
-    &__radio-button {
-        display: none;
-    }
-}
-
-.currency-hr {
-    height: 2px;
-    margin: 1.6rem 0 0;
-    width: 100%;
-    border-top: 1px solid var(--general-section-2);
-}
-
-.currency-selector {
-    &__radio-group {
-        display: flex;
-        flex-direction: column;
-        flex-grow: 1;
-        justify-content: space-around;
-    }
-}
