--- packages/trader/src/sass/details-form.scss
@@ -1,57 +1,0 @@
-@import 'app/_common/base/index';
-
-.details-form {
-    flex-grow: 1;
-
-    & .datepicker {
-        max-width: initial;
-        position: relative;
-
-        &--active-label {
-            label {
-                color: var(--text-general);
-            }
-        }
-        & .icon-datepicker {
-            top: 1.2rem;
-            right: 1rem;
-            position: absolute;
-        }
-        &__picker {
-            margin-top: 8px;
-        }
-    }
-    &__elements {
-        flex-grow: 1;
-        margin: 1.2rem auto;
-        width: 40%;
-
-        &-container {
-            height: calc(100% - 76px);
-            overflow: hidden;
-        }
-        .dc-input {
-            label {
-                top: 1.2rem;
-                background-color: var(--general-main-2);
-            }
-        }
-        .dc-dropdown__display {
-            background-color: var(--general-main-2);
-
-            &-placeholder-text {
-                background-color: var(--general-main-2);
-            }
-        }
-    }
-    &__description {
-        text-align: center;
-        font-size: var(--text-size-xxs);
-        margin: 1.6rem 0;
-        color: var(--text-general);
-    }
-    .address-state__fieldset {
-        margin-bottom: 3.2rem;
-        margin-top: -8px;
-    }
-}
