--- packages/trader/src/sass/terms-of-use.scss
@@ -1,31 +1,0 @@
-@import 'app/_common/base/index';
-
-.terms-of-use {
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-    margin: 0 8rem 0.8rem;
-
-    & .dc-checkbox__label {
-        @include typeface(--paragraph-left-normal-black);
-        text-transform: none;
-        color: var(--text-general);
-    }
-    &__checkbox {
-        margin-top: 2rem;
-    }
-    h4 {
-        @include typeface(--paragraph-left-bold-black);
-        margin: 1rem 0;
-        text-transform: none;
-        color: var(--text-general);
-    }
-    p {
-        @include typeface(--paragraph-left-normal-black);
-        text-transform: none;
-        color: var(--text-general);
-    }
-    input {
-        display: none;
-    }
-}
